login com o facebook

+++ lib/models/product.dart
@@ -6,7 +6,7 @@ import 'package:loja_virtual/models/item_size.dart';
 import 'package:uuid/uuid.dart';
 
 class Product extends ChangeNotifier {
-  Product({this.id, this.name, this.description, this.images, this.sizes}) {
+  Product({this.id, this.name, this.description, this.images, this.deleted = false, this.sizes}) {
     images = images ?? [];
     sizes = sizes ?? [];
   }
@@ -16,12 +16,14 @@ class Product extends ChangeNotifier {
     name = document['name'] as String;
     description = document['description'] as String;
     images = List<String>.from(document.data['images'] as List<dynamic>);
+    deleted = (document.data['deleted'] ?? false) as bool;
     sizes = (document.data['sizes'] as List<dynamic> ?? [])
         .map((s) => ItemSize.fromMap(s as Map<String, dynamic>))
         .toList();
   }
 
   String id;
+  bool deleted;
   String name;
   String description;
   List<String> images;
@@ -34,7 +36,7 @@ class Product extends ChangeNotifier {
 
   bool _loading = false;
   bool get loading => _loading;
-  set loading(bool value){
+  set loading(bool value) {
     _loading = value;
     notifyListeners();
   }
@@ -55,13 +57,13 @@ class Product extends ChangeNotifier {
   }
 
   bool get hasStock {
-    return totalStock > 0;
+    return totalStock > 0 && !deleted;
   }
 
   num get basePrice {
     num lowest = double.infinity;
     for (final size in sizes) {
-      if (size.price < lowest && size.hasStock) lowest = size.price;
+      if (size.price < lowest) lowest = size.price;
     }
     return lowest;
   }
@@ -84,6 +86,7 @@ class Product extends ChangeNotifier {
       'name': name,
       'description': description,
       'sizes': exportSizeList(),
+      'deleted': deleted
     };
     if (id == null) {
       final doc = await firestore.collection('products').add(data);
@@ -103,19 +106,23 @@ class Product extends ChangeNotifier {
         updateImages.add(url);
       }
     }
-    for(final image in images){
-      if(!newImages.contains(image)){
-        try{
+    for (final image in images) {
+      if (!newImages.contains(image) && image.contains('firebase')) {
+        try {
           final ref = await storage.getReferenceFromUrl(image);
           await ref.delete();
-        } catch (e){
+        } catch (e) {
           debugPrint('Falha ao deletar $image');
         }
       }
     }
-    await firestoreRef.updateData({'images' : updateImages});
+    await firestoreRef.updateData({'images': updateImages});
     images = updateImages;
     loading = false;
+  }
+
+  void delete() {
+    firestoreRef.updateData({'deleted': true});
   }
 
   Product clone() {
@@ -125,6 +132,7 @@ class Product extends ChangeNotifier {
       description: description,
       images: List.from(images),
       sizes: sizes.map((size) => size.clone()).toList(),
+      deleted: deleted,
     );
   }
 }
